func/requirements.txt
google-cloud-iam==2.6.1
requests==2.27.1
voluptuous==0.13.1
